requirements.txt
flake8
psycopg2-binary==2.8.4
django-phonenumber-field
phonenumbers
africastalking
retry

python-dotenv

django-referrer-policy

django-feature-policy

Django==2.2.7
